+++ frontend/ui/src/app/components/pages/society/pedigree-tree/family-member-form/family-member-form.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FamilyMemberFormComponent } from './family-member-form.component';
+
+describe('FamilyMemberFormComponent', () => {
+  let component: FamilyMemberFormComponent;
+  let fixture: ComponentFixture<FamilyMemberFormComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [FamilyMemberFormComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(FamilyMemberFormComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});