+++ frontend/ui/src/app/enums/lineage.ts
@@ -3,5 +3,6 @@ export enum Lineage {
   HOUSE = "House",
   CLAN = "Clan",
   FAMILY = "Family",
+  ORPHAN = "Orphan",
   NONE = "None"
 }