+++ frontend/ui/src/app/enums/marital-status.ts
@@ -1,0 +1,7 @@
+export enum MaritalStatus {
+  ALL = "ALL",
+  SINGLE = "SINGLE",
+  MARRIED = "MARRIED",
+  DIVORCED = "DIVORCED",
+  WIDOWED = "WIDOWED",
+}