+++ frontend/ui/src/app/interfaces/models/society/family-member.ts
@@ -2,6 +2,9 @@ export interface FamilyMember {
   name: string;
   born: number;
   died?: number;
-  location: string;
+  location?: string;
+  gender: string;
   parents?: FamilyMember[];
+  spouses?: FamilyMember[];
+  children?: FamilyMember[];
 }