+++ veltahleon/src/main/java/com/testament/veltahleon/enumerations/MaritalStatus.java
@@ -1,0 +1,17 @@
+package com.testament.veltahleon.enumerations;
+
+import lombok.Getter;
+
+@Getter
+public enum MaritalStatus {
+    SINGLE("SINGLE"),
+    MARRIED("MARRIED"),
+    DIVORCED("DIVORCED"),
+    WIDOWED("WIDOWED");
+
+    private final String maritalStatus;
+
+    MaritalStatus(String maritalStatus) {
+        this.maritalStatus = maritalStatus;
+    }
+}