+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/society/PersonDetails.java
@@ -3,6 +3,7 @@ package com.testament.veltahleon.model.entities.society;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.testament.veltahleon.enumerations.ClosedAnswer;
 import com.testament.veltahleon.enumerations.Lineage;
+import com.testament.veltahleon.enumerations.MaritalStatus;
 import com.testament.veltahleon.model.entities.calendar.Year;
 import com.testament.veltahleon.model.entities.dogma.Religion;
 import com.testament.veltahleon.model.entities.places.Nation;
@@ -31,7 +32,7 @@ public class PersonDetails {
             CascadeType.REMOVE
     })
     @JoinColumn(name = "person_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    //@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private Person person;
 
     @Column(name = "birth_day")
@@ -63,7 +64,7 @@ public class PersonDetails {
             CascadeType.MERGE,
             CascadeType.PERSIST})
     @JoinColumn(name = "religion_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    //@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private Religion religion;
 
 //    @OneToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
@@ -92,7 +93,7 @@ public class PersonDetails {
             CascadeType.MERGE,
             CascadeType.PERSIST})
     @JoinColumn(name = "nation_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    //@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private Nation nation;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
@@ -100,7 +101,7 @@ public class PersonDetails {
             CascadeType.MERGE,
             CascadeType.PERSIST})
     @JoinColumn(name = "job_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    //@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private Job job;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
@@ -108,7 +109,7 @@ public class PersonDetails {
             CascadeType.MERGE,
             CascadeType.PERSIST})
     @JoinColumn(name = "rank_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    //@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private Rank rank;
 
 //    @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
@@ -131,8 +132,12 @@ public class PersonDetails {
     @Column(name = "lineage")
     private Lineage lineage;
 
+    @Enumerated(EnumType.STRING)
+    @Column(name = "marital_status")
+    private MaritalStatus maritalStatus;
+
     @Column(name = "biography", columnDefinition = "longtext")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+    //@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private String biography;
 
     @Column(name = "image_URL")