+++ veltahleon/src/main/java/com/testament/veltahleon/rest/society/PersonDetailsController.java
@@ -8,9 +8,14 @@ import com.testament.veltahleon.services.ifc.society.PersonDetailsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.util.Map;
 
@@ -37,6 +42,12 @@ public class PersonDetailsController {
                 .message("Person details retrieved!")
                 .build()
         );
+    }
+
+    @GetMapping(path = "/personDetails/images/{imageName}", produces = MediaType.IMAGE_PNG_VALUE)
+    public byte[] getPersonalImage(@PathVariable("imageName") String imageName) throws IOException {
+        Path image = Paths.get(System.getProperty("user.dir") + "/src/main/resources/assets/images/people/" + imageName);
+        return Files.readAllBytes(image);
     }
 
     @DeleteMapping("/delete/personDetails/{id}")